requirements.txt
streamlit==1.10.0
pandas==1.4.2
Pillow==9.0.1
numpy==1.21.5
seaborn==0.11.2
matplotlib==3.5.1
plotly==5.6.0